Fix reusing a *Tokenizer after an empty stream has been closed
When parsing an empty reader, Stream.head does not get allocated by the pool
at any time and holds a pointer to undefToken.

Previously, on Stream.Close() it was released to the pool, which added the
pointer to undefToken to the pool. As the next call to ParseStream() has to
allocate a new token for the first keyword, it got it from the pool and
received the last released token, which was the pointer to undefToken.
Finally when Stream.IsValid() checked if Stream.head == undefToken, this
evaluated to true, and returned false.

Now, to make sure that we never release a pointer to undefToken into the pool,
we add a check for this in Stream.Close().

Closes: #30